cloneTemplate and removeDeliveryOption
